+++ dsmrreader/config/django_overrides.py
@@ -1,6 +1,6 @@
 """
     Django settings overwritten from env.
-    Mostly try-except because we don't want to pin any defaults to fallback to, just let Django set them.
+    Mostly try-except because we don't want to pin any defaults to fall back to, just let Django set them.
 """
 from decouple import config, UndefinedValueError